+++ CHANGELOG.md
@@ -3,6 +3,12 @@
 * Minor releases (X.1.X) are new features such as added functions or small changes that don't cause major compatibility issues.
 * Major releases (1.X.X) are major new features or changes that break backward compatibility in a big way.
 
+## [Latest](https://github.com/int-brain-lab/iblutil/commits/main) [1.14.0]
+
+### Added
+
+- io.jsonable.load_task_jsonable: read and format iblrig raw data to a trials table Dataframe and a list of raw Bpod trials 
+
 ## [Latest](https://github.com/int-brain-lab/iblutil/commits/main) [1.13.0]
 
 ### Added